requirements.txt
streamlit
pandas
requests
beautifulsoup4
parkrun-api